Added my solution to learnyounode challenge #3
Add file-io.js for solution #3

Hrm, I'd like to dispute the very simple official answer, in that
they always decrement the count of the newlines -- I'd assume that
a text file could be passed that is not newline terminated, and thus,
you have to use conditional logic (a.k.a. an IF) to test if the
very last character of the string from the converted file buffer is
a newline or not.